requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
litellm>=1.0.0
anthropic>=0.40.0
rapidfuzz>=3.0.0
asyncio-throttle>=1.0.0
click>=8.0.0
python-dotenv>=1.0.0
httpx>=0.25.0